server/requirements.txt
flask
flask-cors
flask-socketio
mysql-connector-python
groq
python-dotenv
gunicorn
eventlet
scikit-learn
numpy
